重构监控模型，优化数据提交效率；完善集合服务代码注入与删除功能

+++ jmicro.api/src/main/java/org/jmicro/api/monitor/AbstractMonitorDataSubscriber.java
@@ -5,10 +5,10 @@ import org.jmicro.common.CommonException;
 
 public abstract class AbstractMonitorDataSubscriber implements IMonitorDataSubscriber {
 
-	public static final Integer[] YTPES = ISession.STATIS_TYPES;
+	public static final Short[] YTPES = ISession.STATIS_TYPES;
 	
 	@Override
-	public Double getData(String srvKey,Integer type) {
+	public Double getData(String srvKey,Short type) {
 		throw new CommonException("getData not support for type: "+type);
 	}
